Sync serialization refactor - Move Sync serialization to Common - Move MpReflection to Common - Add simple serialization tests - Make Thing sync worker implicit - Update language version to 12
