+++ Source/Client/Syncing/ApiSerialization.cs
@@ -4,7 +4,7 @@ using Multiplayer.Client.Util;
 
 namespace Multiplayer.Client;
 
-public static class ImplSerialization
+public static class ApiSerialization
 {
     public static Type[] syncSimples;
     public static Type[] sessions;
