+++ Source/Client/Syncing/Dict/SyncDict.cs
@@ -1,17 +1,18 @@
-namespace Multiplayer.Client
-{
-    public static class SyncDict
-    {
-        internal static SyncWorkerDictionaryTree syncWorkers;
+namespace Multiplayer.Client;
 
-        public static void Init()
-        {
-            syncWorkers = SyncWorkerDictionaryTree.Merge(
-                SyncDictMisc.syncWorkers,
-                SyncDictRimWorld.syncWorkers,
-                SyncDictDlc.syncWorkers,
-                SyncDictMultiplayer.syncWorkers
-            );
-        }
+public static class SyncDict
+{
+    internal static SyncWorkerDictionaryTree syncWorkers;
+
+    public static void Init()
+    {
+        syncWorkers = SyncWorkerDictionaryTree.Merge(
+            SyncDictMisc.syncWorkers,
+            SyncDictRimWorld.syncWorkers,
+            SyncDictDlc.syncWorkers,
+            SyncDictMultiplayer.syncWorkers
+        );
+
+        SyncSerialization.syncTree = syncWorkers;
     }
 }
