+++ Source/Common/Syncing/Logger/IHasLogger.cs
