+++ Source/Common/Syncing/Logger/LogNode.cs
@@ -4,12 +4,12 @@ namespace Multiplayer.Client
 {
     public class LogNode
     {
-        public LogNode parent;
+        public LogNode? parent;
         public List<LogNode> children = new();
         public string text;
         public bool expand;
 
-        public LogNode(string text, LogNode parent = null)
+        public LogNode(string text, LogNode? parent = null)
         {
             this.text = text;
             this.parent = parent;
