+++ Source/Common/Syncing/Logger/LoggingByteWriter.cs
