+++ Source/Common/Syncing/SerializationException.cs
