+++ Source/Common/Syncing/Worker/SyncWorkerTypeHelper.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace Multiplayer.Client;
+
+public static class SyncWorkerTypeHelper
+{
+    public new static Func<ushort, Type, Type> GetType = (_, _) => throw new Exception("No implementation");
+    public static Func<Type, Type, ushort> GetTypeIndex = (_, _) => throw new Exception("No implementation");
+}
