+++ Source/Common/Util/CompilerTypes.cs
@@ -16,6 +16,17 @@ namespace System.Runtime.CompilerServices
             BuilderType = builderType;
         }
     }
+
+    [AttributeUsage(AttributeTargets.Class | AttributeTargets.Field | AttributeTargets.Property | AttributeTargets.Struct, Inherited = false)]
+    public sealed class RequiredMemberAttribute : Attribute
+    {
+    }
+
+    [AttributeUsage(System.AttributeTargets.All, AllowMultiple = true, Inherited = false)]
+    public sealed class CompilerFeatureRequiredAttribute : Attribute
+    {
+        public CompilerFeatureRequiredAttribute(string name) { }
+    }
 }
 
 namespace System.Diagnostics.CodeAnalysis
